requirements.txt
crewai==0.95.0
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6 